# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeadMind API",
    "description": "\nRead access to the LeadMind corpus, and the human review queue for duplicates the pipeline\ndeliberately refused to merge.\n\n**Data quality is not lead quality.** Every `quality` field in this API answers *how much do we\nreliably know about this record* \u2014 completeness, reachability, verifiability. Whether a lead is\nworth contacting is a separate judgement with different inputs, and arrives in Phase 5.\n\n**`unknown` is not `unreachable`.** A domain with no MX record is a measurement. A resolver\ntimeout is the absence of one. They are different values everywhere in this API and must not be\ncollapsed by a client.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "List companies",
        "operationId": "list_companies_api_v1_companies_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on domain or name",
              "title": "Q"
            },
            "description": "Substring match on domain or name"
          },
          {
            "name": "min_branches",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "`2` gives the 28 multi-branch companies \u2014 franchises, mostly",
              "title": "Min Branches"
            },
            "description": "`2` gives the 28 multi-branch companies \u2014 franchises, mostly"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "branches, domain, created or quality; prefix `-` for descending",
              "default": "-branches",
              "title": "Sort"
            },
            "description": "branches, domain, created or quality; prefix `-` for descending"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps.",
              "title": "Page Size"
            },
            "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CompanySummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/companies/{company_id}": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "One company and all its branches",
        "operationId": "get_company_api_v1_companies__company_id__get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/meta/locations": {
      "get": {
        "tags": [
          "companies",
          "metadata"
        ],
        "summary": "Resolved locations with live lead counts",
        "description": "Only gazetteer-resolved places appear here.\n\nThe 16 city values that never resolved are kept on their leads verbatim and reported through\n`location.raw` with `resolved: false`, rather than being fuzzy-matched into a plausible wrong\ntown and then offered here as if they were real.",
        "operationId": "list_locations_api_v1_meta_locations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 1000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LocationCount"
                  },
                  "title": "Response List Locations Api V1 Meta Locations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/duplicates": {
      "get": {
        "tags": [
          "review queue"
        ],
        "summary": "List duplicate candidates awaiting review",
        "operationId": "list_duplicates_api_v1_duplicates_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/DuplicateStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Defaults to everything; usually `pending`",
              "title": "Status"
            },
            "description": "Defaults to everything; usually `pending`"
          },
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DuplicateMethod"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "`shared_website` is usually a franchise; `fuzzy_name` is a hypothesis. Exact-key methods never appear here \u2014 they are merged during ingest.",
              "title": "Method"
            },
            "description": "`shared_website` is usually a franchise; `fuzzy_name` is a hypothesis. Exact-key methods never appear here \u2014 they are merged during ingest."
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "name": "max_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Confidence"
            }
          },
          {
            "name": "lead_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only pairs involving this lead",
              "title": "Lead Id"
            },
            "description": "Only pairs involving this lead"
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "confidence, created or method; prefix `-` for descending",
              "default": "-confidence",
              "title": "Sort"
            },
            "description": "confidence, created or method; prefix `-` for descending"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "1-based page number",
              "default": 1,
              "title": "Page"
            },
            "description": "1-based page number"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps.",
              "title": "Page Size"
            },
            "description": "Rows per page. Capped by LEADMIND_API_MAX_PAGE_SIZE \u2014 an unbounded list endpoint is a denial of service with extra steps."
          }
        ],
        "responses": {
          "200": {
            "description": "Pairs, each with both leads embedded and the field diff pre-computed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DuplicateCandidateOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/duplicates/{candidate_id}": {
      "get": {
        "tags": [
          "review queue"
        ],
        "summary": "One pair, side by side",
        "operationId": "get_duplicate_api_v1_duplicates__candidate_id__get",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateCandidateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/duplicates/{candidate_id}/decision": {
      "post": {
        "tags": [
          "review queue"
        ],
        "summary": "Decide a pair",
        "description": "Record a judgement.\n\n`confirmed_duplicate` points the loser at the survivor \u2014 a pointer, not a deletion. Nothing\nis moved or removed, so the decision is fully reversible by sending `pending`, which is what\nmakes it safe to decide quickly.\n\n`rejected_distinct` records that the pair is two real businesses. That answer is as valuable\nas a confirmation: the confirm rate per detection method in `GET /stats/review` is the\ndetector's measured precision, and it is what the fuzzy-name threshold should be tuned\nagainst.",
        "operationId": "decide_duplicate_api_v1_duplicates__candidate_id__decision_post",
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated candidate, and what the decision did to the leads",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DuplicateDecisionResponse"
                }
              }
            }
          },
          "409": {
            "description": "The decision conflicts with a merge already applied to one of these leads. Undo that one first."
          },
          "422": {
            "description": "survivor_id is not one of the two leads in the pair."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "primary_domain": {
            "type": "string",
            "title": "Primary Domain"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "website_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website Url"
          },
          "branch_count": {
            "type": "integer",
            "title": "Branch Count",
            "description": "Leads sharing this domain"
          },
          "locations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locations",
            "description": "Distinct resolved cities across the branches"
          },
          "mean_quality": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Quality"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry",
            "description": "Empty across this dataset. The source workbook contains no firmographics at all; Phase 3 enrichment fills this from the crawl."
          },
          "company_size": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Size"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "leads": {
            "items": {
              "$ref": "#/components/schemas/LeadSummary"
            },
            "type": "array",
            "title": "Leads"
          }
        },
        "type": "object",
        "required": [
          "id",
          "primary_domain",
          "branch_count",
          "created_at"
        ],
        "title": "CompanyDetail"
      },
...
      "DuplicateCandidateOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "method": {
            "$ref": "#/components/schemas/DuplicateMethod"
          },
          "is_auto_mergeable": {
            "type": "boolean",
            "title": "Is Auto Mergeable",
            "description": "False for everything in this queue by construction \u2014 exact-key matches are merged during ingest and never reach a human."
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "status": {
            "$ref": "#/components/schemas/DuplicateStatus"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "lead_a": {
            "$ref": "#/components/schemas/LeadSummary"
          },
          "lead_b": {
            "$ref": "#/components/schemas/LeadSummary"
          },
          "comparison": {
            "items": {
              "$ref": "#/components/schemas/FieldComparison"
            },
            "type": "array",
            "title": "Comparison"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "method",
          "is_auto_mergeable",
          "confidence",
          "status",
          "lead_a",
          "lead_b",
          "created_at"
        ],
        "title": "DuplicateCandidateOut"
      },
      "DuplicateDecisionRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "confirmed_duplicate",
              "rejected_distinct",
              "pending"
            ],
            "title": "Decision"
          },
          "reviewer": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Reviewer",
            "description": "Who decided. Recorded, because an unattributed judgement cannot be audited or reweighted later."
          },
          "survivor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survivor Id",
            "description": "Which of the two leads to keep. Defaults to the better-evidenced one: higher quality score, then more identifiers, then the earlier row."
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "reviewer"
        ],
        "title": "DuplicateDecisionRequest",
        "description": "A reviewer's answer.\n\n``pending`` is a legal decision, not an omission: it is the undo. Setting a decided pair back\nto pending reverses any merge the confirmation applied and returns the pair to the queue."
      },
      "DuplicateDecisionResponse": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/DuplicateCandidateOut"
          },
          "merged_lead_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merged Lead Id",
            "description": "The lead now pointing at a survivor, if the merge was applied"
          },
          "survivor_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Survivor Id"
          },
          "unmerged": {
            "type": "boolean",
            "title": "Unmerged",
            "description": "True when this decision reversed an earlier merge",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "candidate"
        ],
        "title": "DuplicateDecisionResponse"
      },
      "DuplicateMethod": {
        "type": "string",
        "enum": [
          "exact_email",
          "exact_phone",
          "exact_facebook",
          "shared_website",
          "fuzzy_name"
        ],
        "title": "DuplicateMethod",
        "description": "How two records were found to be potentially the same lead.\n\nThe first three are exact and safe to auto-merge. The last two are *not*: a shared website\nis routinely a franchise relationship (five Pumo Technovation branches share one corporate\ndomain and are five real prospects), and a high name similarity is a hypothesis."
      },
      "DuplicateStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed_duplicate",
          "rejected_distinct"
        ],
        "title": "DuplicateStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocationCount": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "lead_count": {
            "type": "integer",
            "title": "Lead Count"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "name",
          "lead_count"
        ],
        "title": "LocationCount"
      },
...
      "Page_CompanySummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CompanySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Rows matching the filter, ignoring pagination"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "description": "Total number of pages at this page size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "Page[CompanySummary]"
      },
      "Page_DuplicateCandidateOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DuplicateCandidateOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Rows matching the filter, ignoring pagination"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page Size"
          },
          "pages": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Pages",
            "description": "Total number of pages at this page size"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages",
          "has_next",
          "has_previous"
        ],
        "title": "Page[DuplicateCandidateOut]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}